+++ packages/gatsby/src/joi-schemas/joi.ts
@@ -1,4 +1,4 @@
-import Joi from "@hapi/joi"
+import Joi from "joi"
 import { IGatsbyConfig, IGatsbyPage, IGatsbyNode } from "../redux/types"
 
 const stripTrailingSlash = (chain: Joi.StringSchema): Joi.StringSchema =>
@@ -86,6 +86,7 @@ export const pageSchema: Joi.ObjectSchema<IGatsbyPage> = Joi.object()
     component: Joi.string().required(),
     componentChunkName: Joi.string().required(),
     context: Joi.object(),
+    // eslint-disable-next-line @typescript-eslint/naming-convention
     pluginCreator___NODE: Joi.string(),
     pluginCreatorId: Joi.string(),
   })