+++ packages/gatsby/src/redux/types.ts
@@ -33,6 +33,7 @@ export interface IGatsbyPage {
   isCreatedByStatefulCreatePages: boolean
   context: Record<string, unknown>
   updatedAt: number
+  // eslint-disable-next-line @typescript-eslint/naming-convention
   pluginCreator___NODE: Identifier
   pluginCreatorId: Identifier
   componentPath: SystemPath